[1] Fixed tool offsets compensation
Previous version of tool offsets compensation was completely wrong due
to lack of understanding of how G92 works.

G92 accepts coordinates and makes it so that the coordinate systems are
offset in such a way that coordinates in the active coordinate system
equal the passed ones. For tool offsets compensation it means that new
coordinates should be calculated based on difference in previous and
current tools offsets and current device coordinates.

Signed-off-by: Innokentiy Alaytsev <alaitsev@gmail.com>