[12] Priority commands are no longer blocked in panic state
From what I can see, most of priority commands should be passed through
when device is in panic state, because these commands can be used to
bring device back to normal state. At very least, this works for GRBL
devices. This is the simplest way of allowing homing after alarm was
triggered. Another approach requires analysing the exact commands and
filtering them based on the knowledge about what commands are allowed to
be executed in panic state.

Signed-off-by: Innokentiy Alaytsev <alaitsev@gmail.com>